requirements.txt
pbr>=1.6
pygerrit>=1.0.0
pyyaml==3.11
six==1.10.0
python-redmine==1.5.0
python-jenkins==0.4.13
requests
